Refactoring @Request scoped bean
